Change pylama, deoplete and neovim integration
Changing PATH and PYTHONPATH to provide support for neovim plugins each
time when activating a virtualenv is not necessary - it can be done just
for running an instance of neovim.

As for ensuring deoplete and pylama see proper packages and executables
when executing neovim: it's not necessary to activate neovim virtualenv
when PYTHONPATH and PATH can be modified locally regardless of whether a
virtualenv is active or not.