Fix configuration of python3 support for neovim
It turns out deoplete didn't require adding site-packages directory
containing jedi python package to PYTHONPATH - adding it to the variable
just fixed an incorrect assignment of python3_host_prog variable. The
assignment was incorrect because it relied on incorrect assumption the
"~" character in the path would be replaced with current home directory.

This commit fixes the path to python interpeter and moves the assignment
before the call enabling deoplete. Since modification of PYTHONPATH is
no longer needed, it is being removed from nvim function in nvim.fish.