Remove nvim.fish
The only responsibility of the function provided by the file was to run
neovim with a locally modified PATH variable. The modification relied on
adding path to /bin directory of neovim virtualenv to it so that pylama
installed in the virtualenv could be called by neovim instance executed
by the function.

The better solution is not to have this function, but instead to install
pylama with its dependencies in a separate virtualenv or in home
directory, and permanently add proper /bin directory to the PATH so that
the same installation of pylama could be used everywhere.